+++ app/models/event.rb
@@ -10,7 +10,6 @@ class Event < ApplicationRecord
   # == Validations =================================================================
   validates :title, presence: true, length: { maximum: 255 }
   validates_boolean :published
-  validates :location, length: { maximum: 255 }
   validate :minimum_one_date
 
   # == Hooks =======================================================================