+++ app/operations/admin/event/create.rb
@@ -3,7 +3,6 @@ module Operations::Admin::Event
     schema do
       opt :event do
         opt :title
-        opt :location
         opt :published
         opt :description
         opt :event_dates_attributes