+++ test/unit/operations/admin/event/create_test.rb
@@ -19,12 +19,12 @@ module Unit
                                                                 title:                  'Testpost',
                                                                 event_dates_attributes: [{
                                                                   start_date: '2022-02-20T14:00',
-                                                                  end_date:   '2022-02-21T15:00'
+                                                                  end_date:   '2022-02-21T15:00',
+                                                                  location:   'Foobar'
                                                                 }]
                                                               }).model.reload
 
               assert_equal 'Testpost', event.title
-              assert event.location.blank?
               assert_not event.published
               assert_nil event.description
               assert_equal 1, event.event_dates.count
@@ -35,17 +35,16 @@ module Unit
             assert_difference '::Event.count', +1 do
               event = ::Operations::Admin::Event::Create.run!(event: {
                                                                 title:                  'Testpost',
-                                                                location:               'ETH Zurich',
                                                                 published:              true,
                                                                 description:            'A fun event',
                                                                 event_dates_attributes: [{
                                                                   start_date: '2022-02-20T14:00',
-                                                                  end_date:   '2022-02-21T15:00'
+                                                                  end_date:   '2022-02-21T15:00',
+                                                                  location:   'ETH Zurich'
                                                                 }]
                                                               }).model.reload
 
               assert_equal 'Testpost', event.title
-              assert_equal 'ETH Zurich', event.location
               assert event.published?
               assert_equal 'A fun event', event.description
               assert_equal 1, event.event_dates.count