+++ test/unit/operations/admin/event/update_test.rb
@@ -8,12 +8,12 @@ module Unit
           setup do
             store :event, ::Operations::Admin::Event::Create.run!(event: {
                                                                     title:                  'Testpost',
-                                                                    location:               'ETH Zurich',
                                                                     published:              true,
                                                                     description:            'A fun event',
                                                                     event_dates_attributes: [{
                                                                       start_date: '2022-02-20T14:00',
-                                                                      end_date:   '2022-02-21T15:00'
+                                                                      end_date:   '2022-02-21T15:00',
+                                                                      location:   'ETH Zurich'
                                                                     }]
                                                                   }).model.reload
           end
@@ -39,7 +39,6 @@ module Unit
 
             # Check that the attrs are still the initial
             assert_equal 'Testpost', event.title
-            assert_equal 'ETH Zurich', event.location
             assert event.published?
             assert_equal 'A fun event', event.description
             assert_equal 1, event.event_dates.count
@@ -47,7 +46,6 @@ module Unit
             # Update the event
             ::Operations::Admin::Event::Update.run!(id: event.id, event: {
                                                       title:       'New post title',
-                                                      location:    'Some other place',
                                                       published:   false,
                                                       description: 'A fun and cool event'
                                                     })
@@ -57,7 +55,6 @@ module Unit
 
             # Check that the attrs have changed
             assert_equal 'New post title', event.title
-            assert_equal 'Some other place', event.location
             assert_not event.published?
             assert_equal 'A fun and cool event', event.description
             assert_equal 1, event.event_dates.count